Fix: 가격 범위 Long

+++ backend/src/main/java/com/ficrew/yourbutler/realestates/application/command/SearchCommand.java
@@ -20,16 +20,16 @@ public class SearchCommand {
     private TradeType tradeType; // 거래 유형: 매매, 월세, 전세
     private List<RoomType> roomType; // 방종류: 아파트, 오피스텔, 단독다가구, 원룸, 빌라연립, 주택(상가, 한옥, ..)
     // 가격
-    private Integer dwMin; // 보증금/전세가 최소
-    private Integer dwMax;
-    private Integer dpMin; // 매매가 최소
-    private Integer dpMax;
-    private Integer rpMin; // 월세
-    private Integer rpMax;
-    private Integer mfMin; // 관리비
-    private Integer mfMax;
-    private Integer rsMin;
-    private Integer rsMax;
+    private Long dwMin; // 보증금/전세가 최소
+    private Long dwMax;
+    private Long dpMin; // 매매가 최소
+    private Long dpMax;
+    private Long rpMin; // 월세
+    private Long rpMax;
+    private Long mfMin; // 관리비
+    private Long mfMax;
+    private Long rsMin;
+    private Long rsMax;
     private Integer uay; // 사용 승인일 기간 (다중택일)
 
 }
